+++ src/Data/MemoryFrame.cpp
@@ -71,6 +71,32 @@ Red::CName MemoryFrame::get_cname(uint32_t p_offset) const {
   return get_value<Red::CName>(p_offset);
 }
 
+float MemoryFrame::get_fixed_point(uint32_t p_offset) const {
+  auto value = get_value<Red::FixedPoint>(p_offset);
+
+  return static_cast<float>(value.Bits) / (2 << 16);
+}
+
+Red::RectF MemoryFrame::get_rect_float(uint32_t p_offset) const {
+  return get_value<Red::RectF>(p_offset);
+}
+
+Red::Point MemoryFrame::get_point(uint32_t p_offset) const {
+  return get_value<Red::Point>(p_offset);
+}
+
+Red::Point3D MemoryFrame::get_point3d(uint32_t p_offset) const {
+  return get_value<Red::Point3D>(p_offset);
+}
+
+Red::Box MemoryFrame::get_box(uint32_t p_offset) const {
+  return get_value<Red::Box>(p_offset);
+}
+
+Red::Quad MemoryFrame::get_quad(uint32_t p_offset) const {
+  return get_value<Red::Quad>(p_offset);
+}
+
 Red::Vector2 MemoryFrame::get_vector2(uint32_t p_offset) const {
   return get_value<Red::Vector2>(p_offset);
 }
@@ -91,12 +117,36 @@ Red::EulerAngles MemoryFrame::get_euler_angles(uint32_t p_offset) const {
   return get_value<Red::EulerAngles>(p_offset);
 }
 
+Red::Matrix MemoryFrame::get_matrix(uint32_t p_offset) const {
+  return get_value<Red::Matrix>(p_offset);
+}
+
+Red::Transform MemoryFrame::get_transform(uint32_t p_offset) const {
+  return get_value<Red::Transform>(p_offset);
+}
+
+Red::QsTransform MemoryFrame::get_qs_transform(uint32_t p_offset) const {
+  return get_value<Red::QsTransform>(p_offset);
+}
+
 Red::WorldPosition MemoryFrame::get_world_position(uint32_t p_offset) const {
   return get_value<Red::WorldPosition>(p_offset);
 }
 
 Red::WorldTransform MemoryFrame::get_world_transform(uint32_t p_offset) const {
   return get_value<Red::WorldTransform>(p_offset);
+}
+
+Red::Color MemoryFrame::get_color(uint32_t p_offset) const {
+  return get_value<Red::Color>(p_offset);
+}
+
+Red::ColorBalance MemoryFrame::get_color_balance(uint32_t p_offset) const {
+  return get_value<Red::ColorBalance>(p_offset);
+}
+
+Red::HDRColor MemoryFrame::get_hdr_color(uint32_t p_offset) const {
+  return get_value<Red::HDRColor>(p_offset);
 }
 
 Red::Handle<MemoryCurveDataFloat> MemoryFrame::get_curve_data_float(